Add a linear controller for the pendulum
This commit closes #2 by adding a linear controller for the pendulum,
applied every 6 frames (0.1 s).  This controller holds the pendulum in
inverted position, but has saturation so that it can't bring the
pendulum upward from too far away.  This saturation is easily removed by
removing the call to clamp, but I feel it makes the demo more
interesting to play with.